# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "ItehProject API documentation",
        "version": "0.0.1",
        "title": "ItehProject API",
        "contact": {},
        "license": {}
    },
    "host": "localhost:9060",
    "basePath": "/",
...
    "paths": {
...
        "/api/_search/users/{query}": {
            "get": {
                "tags": [
                    "user-resource"
                ],
                "summary": "search",
                "operationId": "searchUsingGET",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "query",
                        "in": "path",
                        "description": "query",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/User"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
        "/api/users": {
            "get": {
                "tags": [
                    "user-resource"
                ],
                "summary": "getAllUsers",
                "operationId": "getAllUsersUsingGET",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "userType.in",
                        "in": "query",
                        "required": false,
                        "type": "array",
                        "items": {
                            "type": "string",
                            "enum": [
                                "ST",
                                "LC",
                                "SR",
                                "AD",
                                "SY"
                            ]
                        },
                        "collectionFormat": "multi",
                        "enum": [
                            "ST",
                            "LC",
                            "SR",
                            "AD",
                            "SY"
                        ]
                    },
                    {
                        "name": "userType.equals",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "userType.specified",
                        "in": "query",
                        "required": false,
                        "type": "boolean"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number of the requested page",
                        "required": false,
                        "type": "integer",
                        "format": "int32"
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "description": "Size of a page",
                        "required": false,
                        "type": "integer",
                        "format": "int32"
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sorting criteria in the format: property(,asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
                        "required": false,
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/UserDTO"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "post": {
                "tags": [
                    "user-resource"
                ],
                "summary": "createUser",
                "operationId": "createUserUsingPOST",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "userDTO",
                        "description": "userDTO",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UserDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "user-resource"
                ],
                "summary": "updateUser",
                "operationId": "updateUserUsingPUT",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "userDTO",
                        "description": "userDTO",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UserDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/UserDTO"
                        }
                    },
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/users/authorities": {
            "get": {
                "tags": [
                    "user-resource"
                ],
                "summary": "getAuthorities",
                "operationId": "getAuthoritiesUsingGET",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "*/*"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/users/{login}": {
            "get": {
                "tags": [
                    "user-resource"
                ],
                "summary": "getUser",
                "operationId": "getUserUsingGET",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "login",
                        "in": "path",
                        "description": "login",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/UserDTO"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "user-resource"
                ],
                "summary": "deleteUser",
                "operationId": "deleteUserUsingDELETE",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "login",
                        "in": "path",
                        "description": "login",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "User": {
            "type": "object",
            "required": [
                "activated",
                "fullName",
                "login",
                "userType"
            ],
            "properties": {
                "activated": {
                    "type": "boolean"
                },
                "code": {
                    "type": "string"
                },
                "email": {
                    "type": "string",
                    "minLength": 5,
                    "maxLength": 100
                },
                "firstName": {
                    "type": "string",
                    "minLength": 0,
                    "maxLength": 50
                },
                "fullName": {
                    "type": "string",
                    "minLength": 0,
                    "maxLength": 100
                },
                "id": {
                    "type": "integer",
                    "format": "int64"
                },
                "imageUrl": {
                    "type": "string",
                    "minLength": 0,
                    "maxLength": 256
                },
                "langKey": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 6
                },
                "lastName": {
                    "type": "string",
                    "minLength": 0,
                    "maxLength": 50
                },
                "login": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100,
                    "pattern": "^[_'.@A-Za-z0-9-]*$"
                },
                "resetDate": {
                    "type": "string",
                    "format": "date-time"
                },
                "userType": {
                    "type": "string",
                    "enum": [
                        "ST",
                        "LC",
                        "SR",
                        "AD",
                        "SY"
                    ]
                }
            }
        },
        "UserDTO": {
            "type": "object",
            "required": [
                "login"
            ],
            "properties": {
                "activated": {
                    "type": "boolean"
                },
                "authorities": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "code": {
                    "type": "string"
                },
                "email": {
                    "type": "string",
                    "minLength": 5,
                    "maxLength": 100
                },
                "firstName": {
                    "type": "string",
                    "minLength": 0,
                    "maxLength": 50
                },
                "fullName": {
                    "type": "string"
                },
                "id": {
                    "type": "integer",
                    "format": "int64"
                },
                "imageUrl": {
                    "type": "string",
                    "minLength": 0,
                    "maxLength": 256
                },
                "langKey": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 6
                },
                "lastName": {
                    "type": "string",
                    "minLength": 0,
                    "maxLength": 50
                },
                "login": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100,
                    "pattern": "^[_'.@A-Za-z0-9-]*$"
                },
                "userType": {
                    "type": "string",
                    "enum": [
                        "ST",
                        "LC",
                        "SR",
                        "AD",
                        "SY"
                    ]
                }
            }
        }
    }
}